% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getLickData}
\alias{getLickData}
\title{Get lickometer data}
\usage{
getLickData(df, CS = "Tone")
}
\arguments{
\item{df}{a dataframe corresponding to a .txt file generated by a raw-data export from Graphic State}

\item{CS}{optional label. Defaults to 'Tone'.}
}
\value{
dataframe ready for analysis
}
\description{
This function takes a raw export file from Graphic State and returns a tidy data set ready for analysis.
}
\details{
The clean_data function takes a data frame or table corresponding to a .txt file generated by
a raw-data export from Graphic State.

The function presumes that the lickometer device is registered as
input A1 and that protocol identities exist as follows in the GraphicState database:

 Protocol 1: Habituation

 Protocol 2: Acquisition

 Protocol 3: Single Retrieval Trial

 Protocol 4: Extinction w/17 trials

 Protocol 5: Extinction w/18 trials

 Protocol 6: LTM

The function also presumes the following state logic for all of these protocols:

 State 1 == the ITI

 State 2 == the pre-CS interval (in our programs, the 20 sec preceding each CS when the DVR is turned on)

 State 3 == the CS

 State 4 == the US (only exists in the acquisition protocol)

The function performs the following operations:

 1. Eliminates uninformative variables.

 2. Assigns correct session and trial numbers to each row of data.

 3. If applicable, assigns a 'CS' variable to indicate if the licks are from the ITI (labeled 'Context')
   or the specific stimulus (labeled with the 2nd parameter of the function or 'Tone' by default).

 4. Calculates lick time for each subject aggregated by session, CS, and trial.

 5. Replaces the protocol number with a descriptive label (acquisition )
}
\author{
Jason Shumake
}

