% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTrial}
\alias{getTrial}
\title{The getTrial is a helper function to wideToLong and takes a string corresponding to a column
header in the wide format and returns the trial number embedded in that header.}
\usage{
getTrial(string)
}
\description{
Assumes the header follows a naming convention of 'ProtocolTimeTrial_Session'
where Time == 'Pre' or 'CS' ('Pre' referring to time before the CS and 'CS' referring to time
during the CS) and Trial == the trial number. The _Session field gives the session number
if there is more than one session. If there is only one session it can be omitted.
Example: 'ExtPre13_2' would refer to the pre-CS interval of the 13th trial of the second session of
}

