% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wideToLong}
\alias{wideToLong}
\title{Wide-to-long format conversion}
\usage{
wideToLong(df, CER = "Freezing")
}
\arguments{
\item{df}{data frame of CER measurements in wide format}

\item{CER}{label for the conditioned emotional response. Defaults to 'Freezing'.}
}
\value{
new data frame
}
\description{
Melts data frame with automatic variable creation and labeling given standard column headers
}
\details{
The wideToLong function takes as its input a data frame/table in wide format (one subject per row,
within-subject variables embedded in column names) and the measure of conditioned emotional response
(either an archived 'summary.txt' file from old lickometer data generated by the deprecated
LickometerTrans program, or a spreadsheet of freezing scores). The 'melt' function is used to return a
data frame/table in long format (each within-subject variable explicitly coded in separate columns).
Assumes the header follows a naming convention of 'ProtocolTimeTrial_Session'
where Time == 'Pre' or 'CS' ('Pre' referring to time before the CS and
              'CS' referring to time during the CS)
Values for three variables are extracted from the header:

Session <- a combination of protocol and session

CS <- 'Pre' indicates the CER to context, 'CS' indicates the CER to the stimulus (e.g., tone)

Trial <- the trial number as determined by the helper function 'getTrial'

Example: 'ExtPre13_2' would refer to the pre-CS interval of the 13th trial of the second session of
extinction. The function would label each row of this column as follows:

   Session = 'Extinction 2'

   CS = 'Context'

   Trial = 13
}
\author{
Jason Shumake
}

